Add ehpa_filings dataset download. (#1563)
This adds an `ehpa_filings` dataset download, which HRA/OCJ asked for.  It's suboptimal in that almost all the data fields reflect the *current* values for the user rather than the values at the time they signed their documents.  Actually getting the latter would require parsing the XML answer files of every HPA filing, which would take a very long time.  But I ran this by Johanna and she said it was fine, so we're just going with this for now; I've also mentioned this limitation in the dataset download.