Make Rollbar always report current request info. (#1600)
It looks like #1574 only caused the username to be reported on Rollbar exceptions where Rollbar was explicitly handed a request object.  This is unfortunate, since a lot of our code that logs errors doesn't have direct access to a request object, even though it's frequently called in the context of a request.

A lot of other web frameworks, like Flask, just have a thread-local request object that can be accessed, and Rollbar takes advantage of this by accessing it whenever it needs to make a report.  It can't do this with Django, however, because Django doesn't have a concept of a thread-local request.

But we can add middleware that simulates Flask's behavior.  It's not a good idea for general use, because it breaks Django conventions and assumptions, but purely for the purposes of Rollbar reporting, I think it's a reasonable solution.

So, this is a bit of a hack that implements such middleware and monkeypatches `rollbar.get_request()`--which is always called by Rollbar when it reports anything--to get the Django request of the current thread.

This solution will only work so long as we're not using Django's new asynchronous functionality, which we probably won't be doing for a long time (especially since it's still in-development).