Take HPD registration dates into account when looking up landlord. (#1618)
If there are _multiple_ HPD registrations for the same BBL or BIN, we just return the first one.  Specifically, the HPD registration data for some buildings goes all the way back to the 1990s, and we've been returning some _really_ out-of-date landlords as a result.

This accounts for the problem by sorting our registrations by the last registration date, and secondarily by the registration end date (the second sort likely isn't needed, but I put it in there just in case).