Protect pages behind login. (#1621)
We're getting reports of people somehow visiting pages in logged-out states, which currently is confusing because they can fill out the form but we respond with an inscrutable error message of `You do not have permission to use this form!`.  This protects such pages behind login in the conventional manner (redirect to login, then redirect back to the original page) to hopefully avoid such confusion.