Remove all code relating to the legacy tenant app. (#1631)
Hooray!  This fixes #1625 and removes all our code related to the legacy tenants app, since we decommissioned it on August 3.  In particular, it means the removal of the `legacy_tenants` Django app, and the removal of `pymongo` as a dependency.  It also deprecates the `prefersLegacyApp` GraphQL field (it will always be `null`), because we don't want out-of-date front-ends crashing if they ask for it.