Retry TextIt requests if needed. (#1641)
Fixes #1567.  I originally used Celery's `autoretry_for` keyword argument here but it was impossible to test (or even to test that I wasn't misspelling `autoretry_for`, as renaming the argument to `LOL` resulted in everything still working) so I reverted to doing things the long way around.  I'm still not particularly convinced this is a good test, or that the retry logic will even work in production, but whatever.