Add a form w/ configurable example service instructions. (#1653)
This adds a form at `/ehp/service-instructions-email/example` that allows users to choose a borough and EHP case type, and presents them with the content of the service instructions email that will be sent for such situations.

Also, NYCHA residents need to serve NYCHA via email, but we haven't decided what content to use for that yet. This adds some scaffolding to make it easier to tailor the instructions for NYCHA users, but doesn't actually change the content yet.