Record current user ID in Rollbar. (#1674)
In #1574 we added information about the currently logged-in user on the server-side, but we still have no such information on the client-side.

This fixes things so that any errors reported from the browser are associated with the user ID experiencing them.  I tested it out and it works well.

I also verified that `Rollbar.configure()` "merges" the `payload.person` object with everything else in `payload`, i.e. it doesn't blow away all the other things we configured on `payload` in `project.context_processors.RollbarSnippet` on the server-side.