Allow norent letters to span multiple rent periods in db. (#1683)
This replaces the `rent_period` foreign key field in norent's `Letter` model with a `rent_periods` many-to-many field, so that we can eventually allow users to send letters that span multiple rent periods.

Note that this _reduces_ the constraints at the database level; it's now technically possible for NoRent letters to exist that have no rent periods associated with them, or multiple letters for the same user and the same rent period, and so on.  The way we're validating things in the application ensures such situations should never actually occur, though.  It's also not so bad to relax the constraints a bit, since there are potential edge cases, e.g. a user who might have emailed their letter in early October but decided to send another via USPS a few days later.

Note also that the data migration included in this PR should be given ample time to execute, since it iterates through all the NoRent letters and migrates them each individually.  We'll need to take the site down during this period, so it will need to be done during off-hours.