Add 2019 Rent Stab data to Data Driven Onboarding (#1690)
This PR updates our DDO menu with the fresh 2019 rent stab data. In doing so, it creates two new graphql fields: stabilizedUnitCount and stabilizedUnitCountYear which we can use to store the year of and count for the most recent rent stabilized unit count data we have available. This change deprecates the stabilizedUnitCount2017 field, as the year suffix is now abstracted away by this new structure.