Add forms for setting NoRent letter rent periods (#1686)
This adds models, forms, and GraphQL endpoints for setting what rent periods a user wants their next NoRent letter to cover.

Note that until we actually use this in the front-end, I've added the new GraphQL session fields under `ignoreFields` in `autogen-config.toml`, to ensure that we don't have to deal with migrations if we decide to change the schema for those fields.  Once we want to use them in the front-end, we should remove them from `ignoreFields`.