Add onboardingInfo.isInLosAngeles. (#1702)
This adds a `session.onboardingInfo.isInLosAngeles` GraphQL field, which returns whether the currently logged-in user is in LA County.  (Note that we already have a `norentScaffolding.isInLosAngeles` field, but it only tells us whether the user currently onboarding to NoRent.org is in LA County, not the currently logged-in user.)

We will need this because we'll soon be providing different SAJE-related functionality to logged-in users who are in LA County.