Send an email to staff when we need to mail an LOC. (#1712)
This adds a new `LOC_EMAIL` setting which, if defined, will be emailed whenever someone sends a Letter of Complaint and asks us to mail it for them.

The email will include information about the LOC's custom issues (if any) and whether the user's landlord details were looked up from open data, which are the core determinants of figuring out whether the letter needs to be investigated for potential follow-up with the user. It also includes links to the admin for mailing the user's letter, making/requesting changes to it, or rejecting it.  Finally, it also sets the "reply to" header of the email to the user's email, so that staff can just reply to the email to start a conversation with the letter's sender.