Support data-jf-show-href-only-in-plaintext. (#1713)
Sometimes we have text like "visit boop.com", where "boop.com" is hyperlinked to https://boop.com.  In this case, when we convert the HTML to plaintext, we just want the plaintext to be "visit https://boop.com".

This adds support for a new attribute called `data-jf-show-href-only-in-plaintext` which, if added to an `<a>`, will tell our html-to-text processor to ignore the link text when converting to plaintext, and just show the href attribute as-is.