Email password reset verification code to user. (#1706)
We've had situations where users with land-lines need to reset their password, or where we're unable to send texts to the user because we get "destination handset unreachable" errors from Twilio.

This change additionally sends the verification code to the user over email, if the user has an email account with us, which is a quick fix to work around the issue.

The "if you didn't receive a code" text has been slightly changed to reflect this alteration. Note that we're vague about it because we don't want to leak information about whether the phone number is associated with an email address (since we only require email addresses recently).