Add California-specific AB-3088 changes to NoRent (#1710)
State bill AB-3088 has passed in California, which requires a specific letter template, which this implements. It also changes the KYR interstitial for California to point to AB3088 resources, and it reconfigures the Los Angeles redirect to be a LA-specific KYR screen that allows the user to opt-in to SAJE communications.