Improve links to static pages from /dev. (#1732)
Until now, one annoying aspect of following links to static pages on `/dev` was that they would immediately redirect to a version of their URL with `?staticView` appended, which was annoying.

I realized, though, that all our static pages have file extensions: `.txt`, `.html`, or `.pdf`.  So this adds a new `isStaticPageRoute()` helper and uses it to render those links as actual `<a>` elements in `/dev` rather than as `<Link>`s, to ensure that navigating to them is never pushState-based.