Allow landlord name to be blank in admin. (#1756)
The `LandlordDetails` model's `name` field wasn't allowed to be blank, yet when we lazily create it in our GraphQL endpoint, we sometimes create it with a blank name (if we couldn't look it up in open data).  This results in an annoying situation where an admin might change an unrelated field in the user's admin, but then get an error message saying that the name of the landlord is blank.

This fixes things so that the admin doesn't mind if the name is blank.  The one GraphQL endpoint that accepts the name requires it to be non-blank, so it shouldn't change anything for users.

Finally, since this change doesn't actually affect the back-end database in any way, I changed an old migration to ensure this minor change doesn't create a pointless migration.