Add ManagementCompanyDetails model (#1755)
This adds a model and admin UI that allows management company details to be manually specified for HP actions.  They are only used if the users' `LandlordDetails.is_looked_up` is `False`.  In other words, it's only possible to manually specify a management company if the user _also_ manually specifies a landlord.  I did this because it was the easiest way to implement this change, but we can revisit it later.

This also adds some documentation to the `LandlordDetails` model in the admin, to explain the importance of the `is looked up` field.