Change EHPA issues and merge them. (#1771)
This adds the ability for multiple short issues to be "merged" into a single issue in the EHPA forms, e.g. rather than listing "no heat" and "no hot water" as separate issue, "no heat & no hot water" are listed as a single issue.  This helps prevent addendumification.

We also change the list of EHPA issues to be only the most common winter season issues.

Finally, it's now possible to enter 6 custom issues per area (instead of 3).  This applies to LOC and EHPA.