Update Evictions DDO card to link to EFNYC (#1773)
Now that we have an updated landing page for EFNYC ( JustFixNYC/eviction-free-nyc#132 ), this PR reconfigures our "Respond to an Eviciton" card on DDO to link to EFNYC instead of the RTC Moratorium FAQs. In doing so, it also adds a new function called efnycURL that adds the correctly formatted locale to our outbound link.