Truncate prior cases in EHPA, if needed. (#1791)
So it turns out that #1590 didn't completely fix #1589, because we just had someone with tons of prior cases still generate an addendum.

This fixes the problem by simply writing "and N more" if there are more than a certain number of cases. The end result is that not all the prior cases are actually listed, which is a bummer, but at least we're not forcing the user to cull their own data. There doesn't seem to be any other way around this, alas.