Add sub-headings to sections of issue checklist (#1775)
This breaks up the "wall of checkboxes" in the issue checklist by breaking it up into distinct sections with headings.

Currently it only does this on the "Bathrooms issues" page, as this is the only page with issues of the form "{category}: {problem}".  It's understood that doing this to _only_ this section is a bit confusing, but we think the benefit outweighs the costs.