Don't log 'you have not chosen rent periods' as a server-side error. (#1798)
We've been having a number of users that run into this error, somehow--perhaps they are using the "back" button on their browser after sending their first letter or something, we'll have to look into analytics--but regardless, they all seem to eventually find their way. So let's stop spamming our error logging service by treating this as the form validation error it is.