Rename routes.ts to route-info.ts, steps.tsx to routes.tsx. (#1796)
It'd be a lot easier to understand our codebase if the definitions of the route info (e.g. the URL paths) were right next to the components that defined the actual React components for those routes.

So, this PR does this by:

* Renaming all `routes.ts` files to `route-info.ts` (this makes more sense anyways since their exports are usually functions of the form `createXXXRouteInfo()`).

* Renaming all `steps.tsx` files, and other files that define primary routes, to `routes.tsx` (this makes more sense anyways since their exports are usually components of the form `XXXRoutes`).

This way every directory containing routes will have `route-info.ts` and `routes.tsx` files in them, and they'll always be right next to each other, and therefore (hopefully) easier for developers to understand the relationship between and locate.