Add better CSS for eviction free site. (#1804)
This adds extremely NoRent-inspired styling to the evictionfree site.

Right now we're largely just importing SCSS directly from the norent directory, which is something we're not doing with our TypeScript code. Until such time that we decide that we really want the site to be styled significantly differently from NoRent, though, this seems like a reasonable decision.

The only content from NoRent's SCSS that has been copy-pasted to the new site are the root styles.scss and _bulma-overrides.scss. This also separates out a bunch of the inline SCSS in norent/styles.scss to a separate norent/_misc.scss file to keep things a bit more DRY.

Also, $justfix-green in the new site is currently set to pink; this is just a placeholder color for now, we'll change it to something more appropriate once Tahnee decides on a real replacement.