Add About and FAQs page with placeholder content to Eviction Free NY (#1817)
This PR adds the About and FAQs pages with whatever content we have so far to populate them with. In doing so, this PR also makes a full copy of all of the NoRent.org SCSS and font files, with the idea that we will trim down what we don't need once the design and styling is finalized.