Speed up PDF overlay for blank overlay pages. (#1820)
The overlaying of PDFs takes longer than I'd like; each page takes about a second on my brand-new computer, and it takes even longer on the staging server.  This speeds things up by not bothering to merge pages whose overlay is blank (which for our purposes means the very first page of the 3-page declaration).