Add preliminary content to Eviction Free confirmation page (#1818)
This PR adds in content and styling for the Eviction Free declaration builder confirmation page. Note: there are a couple "to do" items commented out so that we remember to change some pieces of the content dynamically based on user choices from previous form questions— which will come once access to that user data is set up.