Constrain evictionfree state to NY. (#1825)
This constrains the choice of city/state in evictionfree to NY.

Note that one big downside of this that I'm noticing, which is probably just a less-obvious edge case in the national version, is that Mapbox is suggesting some _neighborhoods_ within NYC, such as Jamaica Estates and Brooklyn Heights, as search results. Because our existing "is this city in NYC" logic only tests against common borough names, we're going to classify such submissions as being "not in NYC", which is not good.