Add scaffolding of Eviction Free landing page (#1829)
This PR is a continuation of the work on #1828: it implements the rest of the structure of the landing page on the Eviction Free site. Content is still placeholder as the actual content for these primary pages is yet to be finalized. Included in here is also some quick fixes to the header that were made clearer through the implementation of the page.