Internationalize the Eviction Free declaration builder flow (#1845)
This PR internationalizes all snippets of text in the Eviction Free declaration builder flow. In doing so, it also updates some content to be in sync with edits in our priority guide. It also internationalizes some components outside of our evictionfree frontend directory, namely landlord-names-and-contact-types.tsx and landlord.tsx which are both used in the landlord name step of the flow.