Internationalize FAQs content for Eviction Free NY (#1853)
This PR internationalizes our FAQs content for the Eviction Free NY site. In doing so, it adds the remaining faq questions, and also refactors some of our structuring to simplify things a bit— namely, it removes the notion of a "preview" and "full" answer that are different from one another.