Add backend logic for evictionfree checkbox forms. (#1857)
This adds validation to the evictionfree checkbox forms for agreeing to legal terms, and asserting that the user is telling the truth. Right now we're not actually saving this information anywhere, meaning the user may have to re-check all the boxes if they ever go back steps, and technically an extremely crafty user could go through the process without technically agreeing to anything (although we can verify that a user successfully submitted these forms by parsing logs)... we can deal with these limitations later though.