Add ArchivedLetterRequest model. (#1926)
Right now, whenever we reject a LOC, or whenever a user emails us and tells us they want to send a new one, we need to delete the user's `LetterRequest` object so they can make a new one.  This isn't ideal because it deletes the information we had about the user's previous letter request, which is particularly bad if the previous request was a real letter that got sent.

While we could make `LetterRequest` have a many-to-one relationship with the user model instead of a one-to-one, this would require us to change a _lot_ of existing code that assumes it's one-to-one.  This PR attempts to simplify the approach by adding a nearly identical model called `ArchivedLetterRequest` instead.  The model has some additional fields for book-keeping, like a "notes" field for documenting why the letter request was archived.

Enough information is stored that, if we later decide that a many-to-one `LetterRequest` is a better idea in the long run, we can make a data migration that converts our `ArchivedLetterRequest` models back into `LetterRequest` models.

This PR also makes certain admin tasks easier by adding the following actions to the letter request inline admin:

* A "Reject letter..." button now takes the user to a page where they can reject the letter, and provide a rejection reason and additional notes.  This ultimately archives the letter, allowing the user to submit a new one.  The action can only be accessed if the letter hasn't already been sent.

* An "Archive letter..." button now takes the user to a page where they can archive the letter, and provide additional notes. This also ultimately archives the letter, allowing the user to submit a new one.  This is essentially an alternative to deleting the letter request outright, and as such, Outreach Coordinators have been stripped of the ability to actually delete letter requests.

The email sent to staff when a letter is submitted for us to send has also been changed to reflect these changes.

Finally, note that while this PR makes some apparent changes to GraphQL files, they're just moving lines around, rather than actually making any meaningful changes.  This is happening because an abstract `BaseLetterRequest` was factored out of `LetterRequest`, which caused some fields to be slightly re-ordered.