Add phone number lookup, search form to Front plugin (#1928)
This adds support for SMS conversations to the Front plugin.  Also, when users can't be found by looking at the recipient, a search form is provided so the user can look up the user manually.

This also adds a bit of support for using HTTPS while `DEBUG` is true: if `SESSION_COOKIE_SECURE` is also set, that means that the site must be served over HTTPS in order to work properly, so the default protocol is HTTPS instead of HTTP (and thus links generated by the server are always HTTPS).