Always fill out initials in HPA packet (#1938)
In the "Tenant repairs allegations MC", which fills out a section where the tenant initials one of three options, we weren't always choosing an option, which was resulting in the court rejecting some of our filed EHPs. Now we always fill out the third option in cases where we previously left them all blank.