Added support banner to FAQs and About pages on EFNY (#1941)
This PR adds a new banner at the bottom of the FAQs and About pages on Eviction Free NY that directs folks to Housing Court Answers if they need additional support. The text is identical to a section we show on the confirmation page, but since the banner styling and html structure was quite different, I chose to create a new component for the banner instead of refactoring the section on the confirmation page— that being said, we are reusing translations, which is good!