Skip asking about tenant children for now. (#1943)
The HPD inspection form has changed. It's no longer asking the applicant to list all their kids; it's now asking for their youngest, and only if the kid either lives there or spends more than 10 hours a week there.

We don't have time to change our end of the flow right now, so for the time being we're just going to skip asking the user about kids at all, since it's a rather invasive question anyways. We'll loop back once we have time to implement the changes to our flow properly.