Render declaration hardship PDFs faster (#1947)
Right now rendering hardship declaration PDFs takes an excessive amount of time (several seconds on our Heroku dynos), and it's largely because PyPDF2 is extremely inefficient in the way it merges pages, often encoding and decoding the same PDF content multiple times per page merge.

This adds some caching that improves the performance of cold renders by about 2x, and reduces subsequent renders by a lot more, since it's able to reuse parsed content stream data from the blank hardship declaration PDFs, which is where the vast majority of time is spent.

One downside of this approach is that we're permanently caching the blank hardship declaration PDFs for all our locales. This isn't a big deal right now, but could become burdensome as we grow the tenant platform--especially if we start using this approach for more forms.