Store GeoJSON point for user address. (#1949)
This stores the user's latitude and longitude as a GeoJSON point when possible in `OnboardingInfo.geometry`.  Note that we're not currently using a GeoDjango `PointField` because we're not actually requiring PostGIS just yet (it's not installed on our staging/prod servers).  If we decide a `PointField` is better, we can always implement a data migration later.

This also adds a `geocoded_address` field to `OnboardingInfo`, similar to the one in the `findhelp` app, that should make it easier to tell exactly what address our derived fields are being calculated from.

Both new fields show up in the Django admin as read-only.