Add account settings page w/ name field. (#1954)
This adds an "account settings" page with a single editable field for the user's name. (It's not linked-to from anywhere, since we don't want to publicly point users to it yet.)

As per the mockups, the user's name is initially read-only, and once the user clicks the "Edit" button, is replaced by a form with "Save" and "Cancel" buttons.  Keyboard focus is maintained throughout, and the "Edit" button has an `aria-label` of "Edit name", to differentiate it from the numerous other edit buttons that will be on the page once we add more settings.