Add email field to account settings page. (#1967)
This adds an email address field to the account settings page. It also adds a new `FormWithRequestMixin` class on the back-end, which the `GrapheneDjangoFormMixin` integrates with, which allows form validation logic to take the current request into account. This new functionality is then used to ensure that everything works OK if the user just submits the form with their current email address (i.e., it ensures that clicking "edit" and then "submit" without changing anything works).