Add a new NYC_GEOSEARCH_ORIGIN setting. (#1964)
Yesterday, the NYC Planning Labs GeoSearch API was down, but there was a staging site available that we could fall back to. Unfortunately, though, we don't currently make it very easy to change the URL of the GeoSearch API, so this adds a new `NYC_GEOSEARCH_ORIGIN` setting that configures things properly on the back-end and front-end.