Require PostGIS and always enable "findhelp" app. (#1982)
In order to support determining what county a user's court is in for EvictionFreeNY.org, we'll need PostGIS.  Since this is necessary functionality, we'll need to _require_ PostGIS.  Since it will no longer be optional, this obviates the need for the `ENABLE_FINDHELP` setting introduced way back in #443.

Also, I had a problem when trying to change the protocol of `DATABASE_URL` on our staging instance--apparently Heroku doesn't like it when we try to change that setting, since it manages the database for us.  I found a [stack overflow post][1] that works around it, but it involved a lot of tomfoolery so I decided to instead just rewrite the Database URL at runtime to use `postgis://` if it was already `postgres://`.  A side benefit of this is that we won't need to change any environment variables on production or staging to have this PR take effect.

Finally, this also runs `manage.py loadfindhelpdata` whenever a new deployment is triggered, and as part of `update.sh`, so that we don't have to worry about manually running it whenever we add/change our built-in shapefile data.

## Breaking changes

Note that something annoying about our `ENABLE_FINDHELP` setting was that **even if findhelp was disabled, the database still regarded its migrations as having been applied**.  This means that all existing local deployments need to run the following command in order to have their findhelp app's migrations actually be copacetic:

    python manage.py migrate findhelp zero --fake

After that, developers should run `bash docker-update.sh` (or `bash update.sh` if they're not using Docker).

[1]: https://stackoverflow.com/a/35704410/2422398