Add geocoded_point field to OnboardingInfo. (#1983)
Since we now require PostGIS, this adds a `geocoded_point` field that represents the user's address as a GeoDjango Point.  This will allow us to easily make geographic queries about the user.

Note that this currently means the user's address will be represented as _both_ a GeoJSON point (see #1949) and a GeoDjango Point.  This is ok for now--we should deprecate the former and use the latter when possible.

This also renders the user's address in the user admin, since we have that functionality from the findhelp app.

In doing so, I fixed a bug whereby our MapBox widget code was out-of-date.  I found some instructions online on how to [fix things][1].

[1]: https://docs.mapbox.com/help/troubleshooting/migrate-legacy-static-tiles-api/#leaflet-implementations