Make national address geocoding more lenient with respect to cities. (#1989)
Sometimes Mapbox gives interesting responses about what city an address is in.

Here's one example: some people claim they live in Wappingers Falls, while Mapbox believes that their street address is actually in Poughkeepsie.  (And interestingly, different map vendors can disagree here; for instance Google Maps might think their address _is_ in Wappingers Falls.)  Because our logic currently filters out results that aren't in the exact city the user said they live in, this has resulted in us believing that the user entered an invalid address.  We will actually present them with a "your address appears to be invalid" dialog box when they enter their full address, though at least they have the option of ignoring it (partly because we anticipated that our logic might be overly stringent).  However, this also renders us unable to geocode their address to e.g. figure out what county they're in, which we need to do in order to send their declaration to the courts.

So, this PR loosens the restrictions on _requiring_ that an address be in the city a user claims it's in, and instead _prioritizes_ such addresses, but also accepts other address suggestions from Mapbox.

Note that we don't actually _replace_ the city name if Mapbox suggests a different one: for example, if Mapbox thinks a user's address is in Poughkeepsie but the user thinks it's in Wappingers Falls, we trust the user over Mapbox.  This is partly because I think it's better to trust the user's opinion, but also because it'd take extra work to do otherwise.