Make verify_addresses work w/ national addresses. (#1990)
This changes `verify_addresses` to work with national addresses (it previously only worked with NYC ones).

The command also no longer does anything with the `address_verified` field, as per #1991--instead it just searches for addresses that have no geocoding information, and attempts to geocode them, doing so automatically if there's an exact case-insensitive match and asking the user for confirmation otherwise.

Finally, it also hides `address_verified` from the admin.  As such, we could say it tentatively deprecates the use of the field for now.