Send a copy of the EFNY declaration to outside-NYC housing court (#1993)
This sends a copy of the EFNY declaration to the housing court outside of NYC _only_ if we know what county the user is located in, since the state really wants that information right now.

This also adds a `session.onboardingInfo.county` GraphQL field, which is currently only used to render the subject line of these emails; but in the future it can be used by front-end code for any other purpose, if needed.

Note that merging this _will_ cause declarations outside of NYC, for users whose address can be geocoded, to be emailed to the outside-NYC housing court.  It won't currently contain any information about the user's housing court because we aren't asking that information yet (that will come in a future PR).  Better something than nothing for now, though, and this will also allow us to start sending emails for outside-NYC folks who have already used our tool.