Make process_declarations send emails to housing court. (#1994)
We now send emails to outside-NYC housing courts (as of #1993), but we need to do this for everyone outside NYC who sent a declaration before we added such functionality.  And because the functionality requires that a user's address be geocoded, we also need to have this ability for the long-term, since it's possible for someone to send a declaration but not actually have their address geocoded until later on.

So this PR adds such functionality to the `process_declarations` command.