Validate ZIP codes by state. (#1995)
Er, so we've actually had a fair number of out-of-state users create accounts on EvictionFreeNY.org claiming that they live in a New York town that has the same name as their out-of-state town, and providing the street address and ZIP code of their out-of-state town.

For instance, we've had at least one instance of someone from Memphis, TN telling us they live in Memphis, New York (an actual town) but providing their TN street address and ZIP code.  Even though Mapbox tells us that the address appears to be invalid, we still allow users to proceed (because sometimes Mapbox is wrong--or at least, sometimes Lob has been wrong and we're assuming that sometimes Mapbox is too).

This attempts to dissuade such use by at least attempting to make sure that the ZIP code the user enters appears to be in New York.  That is, it will reject anything that doesn't either start with a `1` or is one of the handful of ZIP codes in NY that start with `0`.