Add a path to sign in from onboarding. (#1996)
It's currently very difficult for pre-existing users to figure out how to sign in if they followed a link to onboarding.

While there are more high-level UX solutions to this that we should eventually implement, this quick fix shows a notification on the step of onboarding telling users to sign in, if (and only if) they've entered a phone number or email address that's already been taken by an existing account.

Additionally, the EHP splash page has no link for pre-existing users to sign in, like there is on the LOC splash page. So this adds the exact same kind of link to the page.

The only really unfortunate thing now is that once the user signs in, they will likely be taken to the latest step of whatever product they originally signed up for, which isn't particularly useful for e.g. LOC users who now want to file an EHP.  Only the most observant users might discover the "Take action" navbar link, which is currently the main way of locating actions other than the one a user signed up for.  This will probably have to wait to be addressed by our "dashboardification" product improvement.