Make verify_addresses convert NYC-based national addresses. (#2004)
This adds functionality to the `verify_addresses` command to help us convert national addresses that are actually in NYC into proper NYC addresses (see #2002 for details on how to stop such addresses from getting into our database in the first place).

Figuring out how to do this with a minimal amount of new code was actually kind of fun, since conversion requires figuring out what borough the user's address is in.  Instead of digging through the return value Mapbox gives us, which would involve a lot of potentially error-prone new code, I used our new county-finding code introduced in #1985 to look up the county for the geocoded national address and map it to a borough.  This is used both to figure out if the national address is in NYC in the first place, and to determine the address' borough.