Allow users to be impersonated by staff. (#1977)
As we continue to add capabilities to the site it may become convenient for administrators to impersonate users. For example, it's now much more straightforward to change a user's NYC address in their account settings page than it is via the somewhat confusing Django admin UI.

This kind of feature could also be useful for tenant advocates, if we ever want to bring back the concept of an "advocate dashboard", which could allow advocates to use our tools on behalf of the tenants they represent.

It will also be useful for developers during development on their local machines, as it allows them to easily see what the site looks like from any user's perspective without needing to remember any additional phone numbers or passwords.

So, this fixes #967 by adding an impersonation button on all user admin pages.

Clicking the button takes the user to a page that explains what's about to happen, and provides a confirmation button.  

Once the impersonation has begun, the user is effectively logged-in as the user they are impersonating.  Ending the impersonation can be done by going to the site's "log out" page, which shows an additional button for ending the impersonation.

## Security notes

In order to impersonate, a user must be a staff member and have the new `users.impersonate_users` permission, which is automatically given to Outreach Coordinators.  Additionally, staff members can't impersonate other staff members: only superusers can impersonate other staff members.