Reject NYC addresses in national address form. (#2002)
We've had a (relatively small) number of cases in which users enter the name of their _street_ in the _city_ field of the EFNY flow.

For example, if the user lives at 1200 Stratford Ave in the Bronx, they might mistakenly enter "Stratford, NY" when we ask them for their city and state.  This is unintentionally abetted by Mapbox autocomplete and validated by our server, as Stratford is an actual city in New York.

The user is then taken to the _national_ address entry field, since we assume they live outside NYC.  Then they enter the rest of their address, which we geocode to be in NYC.  The problem is that their account will now consist of a NYC address, but our platform will treat them as being a non-NYC user, which is Bad.  For example, in EFNY it means we'll be emailing their declaration to the "outside NY" housing court, and we won't be able to automatically look up their landlord for them.

This adds a quick, albeit not terribly user-friendly fix: when the user enters a street address that geocodes to be within NYC, we reject them with the following error:

> Your address appears to be within New York City. Please go back and enter "New York City" as your city.

Not _too_ helpful but it prevents users from continuing the process with a completely messed up account.  It's also intended as a "last resort": we should also provide a good user experience to ensure that the user knows exactly what information they're submitting to our site, to ensure that this validation code is triggered as rarely as possible.  See #2003 for an example of such UX.